+++ FBSDKCoreKit/FBSDKCoreKit/include/FBSDKInternalUtility.h
@@ -20,17 +20,6 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
-FOUNDATION_EXPORT NSString *const FBSDK_CANOPENURL_FACEBOOK
-  DEPRECATED_MSG_ATTRIBUTE("`FBSDK_CANOPENURL_FACEBOOK` is deprecated and will be removed in the next major release; please use `URLScheme.facebookAPI` instead");
-FOUNDATION_EXPORT NSString *const FBSDK_CANOPENURL_FBAPI
-  DEPRECATED_MSG_ATTRIBUTE("`FBSDK_CANOPENURL_FBAPI` is deprecated and will be removed in the next major release; please use `URLScheme.facebookAPI` instead");
-FOUNDATION_EXPORT NSString *const FBSDK_CANOPENURL_MESSENGER
-  DEPRECATED_MSG_ATTRIBUTE("`FBSDK_CANOPENURL_MESSENGER` is deprecated and will be removed in the next major release; please use `URLScheme.messengerApp` instead");
-FOUNDATION_EXPORT NSString *const FBSDK_CANOPENURL_MSQRD_PLAYER
-  DEPRECATED_MSG_ATTRIBUTE("`FBSDK_CANOPENURL_MSQRD_PLAYER` is deprecated and will be removed in the next major release");
-FOUNDATION_EXPORT NSString *const FBSDK_CANOPENURL_SHARE_EXTENSION
-  DEPRECATED_MSG_ATTRIBUTE("`FBSDK_CANOPENURL_SHARE_EXTENSION` is deprecated and will be removed in the next major release; please use `URLScheme.facebookAPI`");
-
 NS_SWIFT_NAME(InternalUtility)
 @interface FBSDKInternalUtility : NSObject
 #if !TARGET_OS_TV